+++ src/components/departure-list/utils.ts
